reWrite.py is now main.py
removed previous main.py
